+++ R/UseTemplate.R
@@ -1,8 +1,8 @@
 UseTemplateList = function(newfile, fileList, find=NULL, replace=NULL){
-for(i in fileList){
-cat(UseTemplate(i, find=find, replace=replace), file=newfile, append=TRUE)
+for(Template in fileList){
+cat(UseTemplate(Template, find=find, replace=replace), file=newfile, append=TRUE)
 }
-return(invisible(NULL))
+return(invisible(TRUE))
 }
 
 
@@ -28,3 +28,29 @@ cat(OutText, file=NewFile)
 return(OutText)
 }
 
+.ChooseTemplate = function(){
+      if (interactive()) {
+Templates = read.csv(system.file("Templates/templates.csv", package="BrailleR"))
+Types = as.character(unique(Templates$Type))
+        UserPrefType = menu(Types,
+                 title = "Which type of template do you want?")
+Subset = Templates[Templates$Type == Types[UserPrefType], ]
+        UserPrefFile = menu(Subset$Template,
+                 title = "Which template do you want?")
+file = Subset$Template[UserPrefFile]
+message(paste("The", file, "uses the following terms that must be replaced:", Subset$Changes[UserPrefFile], "\nIt will ", Subset$Description[UserPrefFile]))
+      } else {
+        .InteractiveOnly()
+file=NULL
+      }
+return(file)
+}
+
+.BuildRmdFile = function(TemplateList = c("simpleYAMLHeader.Rmd", "StandardSetupChunk.Rmd", "GetLibs.Rmd", "GetData.Rmd"), 
+      Changes = list(), Outfile = "test.Rmd"){
+for(Template in TemplateList){
+cat(.UseTemplate(Template, Changes=Changes), "\n\n\n", file=Outfile, append=TRUE)
+}
+.NewFile(Outfile)
+return(invisible(TRUE))
+}